% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_density.R
\name{add_hr}
\alias{add_hr}
\title{Add home range identifiers}
\usage{
add_hr(net, perc = c(0.5, 0.9, 0.95))
}
\arguments{
\item{net}{The glinet object.}

\item{perc}{Numeric vector of volume percentages to extract home ranges from the underlying utilization
distribution}
}
\value{
The updated glinet object.
}
\description{
This function adds home range identifiers to glinet network vertices.
}
\author{
Vojta Barták
}
