% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_utils.R
\name{get_vertex_weights}
\alias{get_vertex_weights}
\title{Get the vertex weights}
\usage{
get_vertex_weights(net)
}
\arguments{
\item{net}{glinet object}
}
\value{
numeric vector of vertex weights
}
\description{
This function extracts the vertex weights as a numeric vector.
}
\author{
Vojta Barták
}
