% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_export.R
\name{glinet_to_hr_lines}
\alias{glinet_to_hr_lines}
\title{Extract home ranges}
\usage{
glinet_to_hr_lines(net, perc = 0.95)
}
\arguments{
\item{net}{The glinet object.}

\item{perc}{The volume percentage of the underlying utilization distribution that defines home ranges.}
}
\value{
A SpatialLines (sp) object.
}
\description{
This function extract home ranges from the glinet network in the form of spatial lines
each representing one home range.
}
\author{
Vojta Barták
}
