% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_export.R
\name{glinet_to_sp}
\alias{glinet_to_sp}
\title{Convert the network to Spatial Points}
\usage{
glinet_to_sp(net)
}
\arguments{
\item{net}{The glinet object.}
}
\value{
The SpatialPointsDataFrame object.
}
\description{
This function converts the glinet network into SpatialPointsDataFrame with weights in a field "w" and
eventual home range identifiers in appropriately named fields.
}
\author{
Vojta Barták
}
