% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_density.R
\name{heat_glinet}
\alias{heat_glinet}
\title{Simulate the heat equation}
\usage{
heat_glinet(net, sigma, beta = 0.5)
}
\arguments{
\item{net}{A glinet object.}

\item{sigma}{...}

\item{beta}{...}
}
\value{
Returns a new glinet object with vertex weights updated by heat propagation.
}
\description{
This function simulates the heat equation on a linear network discretized into segments of same lengths.
It thus effectively estimates a density of point events (with weights) located on the network vertices.
}
\references{
...
}
\author{
Vojta Barták
}
