% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_export.R
\name{hr_count}
\alias{hr_count}
\title{Determine the number of home ranges}
\usage{
hr_count(net, perc = 0.95, split = TRUE, hr_size = 2000)
}
\arguments{
\item{net}{The glinet object.}

\item{perc}{The volume percentage of the underlying utilization distribution that defines home ranges.}

\item{split}{If TRUE (default), then home ranges longer than 2*hr_size are splitted.}

\item{hr_size}{Length of home range size used for splitting longer home ranges. Not used when split
is FALSE.}
}
\value{
The number of home ranges in the network (integer).
}
\description{
This function determines the number of home ranges from the linear network with
vertex weights representing utilization distribution.
}
\author{
Vojta Barták
}
