% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glinet_density.R
\name{init_glinet}
\alias{init_glinet}
\title{Initialize the network}
\usage{
init_glinet(net, xy, w)
}
\arguments{
\item{net}{A glinet object.}

\item{xy}{A table-like object (data.frame or matrix) with two columns for x and y coordinates,
representing the set of points (observations or events) whose density should be estimated on the
network.}

\item{w}{A numeric vector (of the same length as the number of points in xy) of weights. Each weight
will be distributed to the nearest two vertices of the network.}
}
\value{
An initialized glinet object.
}
\description{
Initializes the network net by a set of points xy and their corresponding weights w.
}
